refactor: switch to separate `main` and `renderer` exports
