--- src/index.ts
@@ -1,19 +1,0 @@
-export type {
-  CreateForwardingMiddlewareFunction,
-  ForwardedAction,
-  ForwardToMiddlewareOptions,
-} from "./middleware.js";
-export {
-  withRedial,
-  type RedialInit,
-  type RedialMainInit,
-  type RedialRendererInit,
-} from "./withRedial.js";
-export type {
-  AnyAction,
-  AnyState,
-  ElectronApiIn,
-  ElectronMainApi,
-  ElectronRendererApi,
-  ProcessName,
-} from "./types.js";
